+++ src/components/dictionaries/deeplx/_locales.ts
@@ -1,0 +1,7 @@
+import { getMachineLocales } from '../locales'
+
+export const locales = getMachineLocales({
+  en: 'DeepLX',
+  'zh-CN': 'DeepLX',
+  'zh-TW': 'DeepLX'
+})