+++ src/components/dictionaries/deeplx/auth.ts
@@ -1,0 +1,6 @@
+export const auth = {
+  apiUrl: '',
+  token: ''
+}
+
+export const url = 'https://github.com/OwO-Network/DeepLX'