+++ src/components/dictionaries/deeplx/config.ts
@@ -1,0 +1,21 @@
+import {
+  MachineDictItem,
+  machineConfig
+} from '@/components/MachineTrans/engine'
+import { Language } from '@opentranslate/translator'
+import { Subunion } from '@/typings/helpers'
+
+export type DeepLXLanguage = Subunion<
+  Language,
+  'zh-CN' | 'zh-TW' | 'en' | 'ja' | 'ko' | 'fr' | 'de' | 'es' | 'ru'
+>
+
+export type DeepLXConfig = MachineDictItem<DeepLXLanguage>
+
+export default (): DeepLXConfig =>
+  machineConfig<DeepLXConfig>(
+    ['zh-CN', 'zh-TW', 'en', 'ja', 'ko', 'fr', 'de', 'es', 'ru'],
+    {},
+    {},
+    {}
+  )